Moved `Variable` expression out of the `Unary` namespace (it's not unary, it doesn't take a single expression as input)
