+++ Yolol.Analysis/TreeVisitor/BaseTreeVisitor.cs
@@ -8,7 +8,7 @@ using Yolol.Grammar.AST.Expressions;
 using Yolol.Grammar.AST.Expressions.Binary;
 using Yolol.Grammar.AST.Expressions.Unary;
 using Yolol.Grammar.AST.Statements;
-using Variable = Yolol.Grammar.AST.Expressions.Unary.Variable;
+using Variable = Yolol.Grammar.AST.Expressions.Variable;
 
 namespace Yolol.Analysis.TreeVisitor
 {
