+++ Yolol.Analysis/Types/ExpressionTypeInference.cs
@@ -9,7 +9,7 @@ using Yolol.Grammar.AST.Expressions.Binary;
 using Yolol.Grammar.AST.Expressions.Unary;
 
 using Type = Yolol.Execution.Type;
-using Variable = Yolol.Grammar.AST.Expressions.Unary.Variable;
+using Variable = Yolol.Grammar.AST.Expressions.Variable;
 
 namespace Yolol.Analysis.Types
 {
